Struct google_networkconnectivity1_alpha1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Networkconnectivity hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_networkconnectivity1_alpha1 as networkconnectivity1_alpha1;
 
use std::default::Default;
use oauth2;
use networkconnectivity1_alpha1::Networkconnectivity;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Networkconnectivity::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_get(...)`, `locations_global_hubs_create(...)`, `locations_global_hubs_delete(...)`, `locations_global_hubs_get(...)`, `locations_global_hubs_get_iam_policy(...)`, `locations_global_hubs_list(...)`, `locations_global_hubs_patch(...)`, `locations_global_hubs_set_iam_policy(...)`, `locations_global_hubs_test_iam_permissions(...)`, `locations_internal_ranges_get_iam_policy(...)`, `locations_internal_ranges_set_iam_policy(...)`, `locations_internal_ranges_test_iam_permissions(...)`, `locations_list(...)`, `locations_operations_cancel(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_policy_based_routes_get_iam_policy(...)`, `locations_policy_based_routes_set_iam_policy(...)`, `locations_policy_based_routes_test_iam_permissions(...)`, `locations_spokes_create(...)`, `locations_spokes_delete(...)`, `locations_spokes_get(...)`, `locations_spokes_get_iam_policy(...)`, `locations_spokes_list(...)`, `locations_spokes_patch(...)`, `locations_spokes_set_iam_policy(...)` and `locations_spokes_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_global_hubs_create(
&self,
request: Hub,
parent: &str
) -> ProjectLocationGlobalHubCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Hub in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent resource's name of the Hub.

pub fn locations_global_hubs_delete(
&self,
name: &str
) -> ProjectLocationGlobalHubDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Hub.

Arguments

name - Required. The name of the Hub to delete.

pub fn locations_global_hubs_get(
&self,
name: &str
) -> ProjectLocationGlobalHubGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Hub.

Arguments

name - Required. Name of the Hub resource to get.

pub fn locations_global_hubs_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationGlobalHubGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_global_hubs_list(
&self,
parent: &str
) -> ProjectLocationGlobalHubListCall<'a, C>

Create a builder to help you perform the following task:

Lists Hubs in a given project and location.

Arguments

parent - Required. The parent resource's name.

pub fn locations_global_hubs_patch(
&self,
request: Hub,
name: &str
) -> ProjectLocationGlobalHubPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single Hub.

Arguments

request - No description provided.

name - Immutable. The name of a Hub resource.

pub fn locations_global_hubs_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationGlobalHubSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_global_hubs_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationGlobalHubTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_internal_ranges_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationInternalRangeGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_internal_ranges_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationInternalRangeSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_internal_ranges_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationInternalRangeTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_operations_cancel(
&self,
request: GoogleLongrunningCancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_policy_based_routes_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationPolicyBasedRouteGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_policy_based_routes_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationPolicyBasedRouteSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_policy_based_routes_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationPolicyBasedRouteTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_spokes_create(
&self,
request: Spoke,
parent: &str
) -> ProjectLocationSpokeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Spoke in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent's resource name of the Spoke.

pub fn locations_spokes_delete(
&self,
name: &str
) -> ProjectLocationSpokeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Spoke.

Arguments

name - Required. The name of the Spoke to delete.

pub fn locations_spokes_get(
&self,
name: &str
) -> ProjectLocationSpokeGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Spoke.

Arguments

name - Required. The name of Spoke resource.

pub fn locations_spokes_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationSpokeGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_spokes_list(
&self,
parent: &str
) -> ProjectLocationSpokeListCall<'a, C>

Create a builder to help you perform the following task:

Lists Spokes in a given project and location.

Arguments

parent - Required. The parent's resource name.

pub fn locations_spokes_patch(
&self,
request: Spoke,
name: &str
) -> ProjectLocationSpokePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single Spoke.

Arguments

request - No description provided.

name - Immutable. The name of a Spoke resource.

pub fn locations_spokes_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationSpokeSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_spokes_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationSpokeTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.